change modal elements
