+++ frontend/src/components/modal/AddModal.jsx
@@ -1,0 +1,98 @@
+/* eslint-disable react/prop-types */
+import React, { useRef, useEffect } from 'react';
+import { Modal, Button, Form } from 'react-bootstrap';
+import { useDispatch } from 'react-redux';
+import { useFormik } from 'formik';
+import { useTranslation } from 'react-i18next';
+import leoProfanity from 'leo-profanity';
+import useValidationSchemas from '../../validation'
+import { setActiveChannel } from '../../store/slices/chatSlice';
+import { useGetChannelsQuery, useAddChannelMutation } from '../../store/middlewares/index';
+
+const AddModal = ({
+  show,
+  close,
+  createNotify,
+}) => {
+  const inputRef = useRef(null);
+  const dispatch = useDispatch();
+  const { data: channels } = useGetChannelsQuery();
+  const [addChannel] = useAddChannelMutation();
+  const { t } = useTranslation();
+  const { modalShema } = useValidationSchemas();
+
+  const channelsNames = channels.map((channel) => channel.name);
+
+  const formik = useFormik({
+    initialValues: {
+    channelName: '',
+    },
+    validationSchema: modalShema(channelsNames),
+    onSubmit: async (values, { resetForm }) => {
+    const newChannel = {
+      name: leoProfanity.clean(values.channelName),
+    };
+    try {
+      const data = await addChannel(newChannel).unwrap();
+      resetForm();
+      close();
+      if (data) {
+        dispatch(setActiveChannel(data.id));
+      }
+      createNotify();
+    } catch (err) {
+      console.error('Failed to add channel', err);
+    }
+    },
+    validateOnChange: false,
+    validateOnBlur: false,
+  });
+
+  useEffect(() => {
+    if (show !== 'delete') {
+      inputRef.current.focus();
+      inputRef.current.select();
+    }
+  }, [show]);
+
+  return (
+    <>
+      <Modal.Header closeButton>
+        <Modal.Title>{t('channel.add')}</Modal.Title>
+      </Modal.Header>
+      <Modal.Body>
+        <Form onSubmit={formik.handleSubmit}>
+          <Form.Group>
+            <Form.Label htmlFor="name" visuallyHidden>{t('channel.channelName')}</Form.Label>
+            <Form.Control
+              type="text"
+              placeholder={t('channel.newName')}
+              name="channelName"
+              ref={inputRef}
+              value={formik.values.channelName}
+              onChange={formik.handleChange}
+              isInvalid={!!formik.errors.channelName}
+            />
+            <Form.Control.Feedback type="invalid">
+              {formik.errors.channelName}
+            </Form.Control.Feedback>
+          </Form.Group>
+        </Form>
+      </Modal.Body>
+      <Modal.Footer>
+        <div className="mt-2 d-flex justify-content-end">
+          <Button variant="secondary" onClick={close}>
+            {t('modal.cancel')}
+          </Button>
+          <Button variant="primary"
+            onClick={formik.handleSubmit}
+            className="ms-2">
+              {t('modal.submit')}
+          </Button>
+        </div>
+      </Modal.Footer>
+    </>
+);
+};
+
+export default AddModal;
