+++ frontend/src/locales/ru.js
@@ -29,12 +29,12 @@ const ru = {
       addPlaceholder: 'Введите имя канала',
       menu: 'Управление каналом',
       remove: 'Удалить',
-      edit: 'Переименовать',
+      rename: 'Переименовать',
       channels: 'Каналы',
-      created: 'Канал создан',
-      edited: 'Канал переименован',
+      create: 'Канал создан',
+      edit: 'Канал переименован',
+      delete: 'Канал удалён',
       newName: 'Введите новое имя канала',
-      removed: 'Канал удалён',
       channelName: 'Имя канала',
       editChannelName: 'Новое имя канала',
     },
